# Hook vector unrolling in LLVMTileFuseAndVectorizePass.